Put input validations into main.rb
