+++ people_controller.rb
@@ -25,7 +25,7 @@ class PeopleController
 
   def all_people_list
     @people.each_with_index do |person, index|
-      puts "#{index}) [#{person.class}] Age: #{person.age} ID: #{person.id} Name: #{person.name}}"
+      puts "#{index}) [#{person.class}] Age: #{person.age} ID: #{person.id} Name: #{person.name}"
     end
   end
 end
